Perbaikan CRUD pada listapp dan penambahan CRUD pada user

+++ app/Http/Controllers/ListAppController.php
@@ -3,10 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\ListApp;
-use Illuminate\Container\Attributes\Auth;
 use Illuminate\Support\Str;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\DB;
+use SweetAlert2\Laravel\Swal;
 use Illuminate\Support\Facades\File;
 
 class ListAppController extends Controller
@@ -14,13 +13,17 @@ class ListAppController extends Controller
 
     public function index()
     {
+        $apps_menu = 'Kelola Aplikasi';
+        $apps_submenu = 'Daftar Aplikasi';
         $list_apps = ListApp::latest()->get();
-        return view('pages.listapp.index', compact('list_apps'));
+        return view('pages.listapp.index', compact('list_apps', 'apps_menu', 'apps_submenu'));
     }
 
     public function create()
     {
-        return view('pages.listapp.create');
+        $apps_menu = 'Kelola Aplikasi';
+        $apps_submenu = 'Tambah Aplikasi';
+        return view('pages.listapp.create', compact('apps_menu', 'apps_submenu'));
     }
 
     public function store(Request $request)
@@ -56,16 +59,32 @@ class ListAppController extends Controller
         }
 
         $app->save();
-        return redirect()->route('listapp.index')->with('success', 'Aplikasi berhasil ditambahkan.');
+        Swal::fire([
+            'title' => 'Informasi',
+            'text' => 'Aplikasi ' . $app->nama . ' berhasil ditambahkan!',
+            'icon' => 'success',
+            'confirmButtonText' => 'Tutup'
+        ]);
+        return redirect()->route('listapp.index');
     }
 
-    public function edit($id)
+    public function show($slug)
     {
-        $app = ListApp::findOrFail($id);
-        return view('listapp.edit', compact('app'));
+        $apps_menu = 'Kelola Aplikasi';
+        $apps_submenu = 'Detail Aplikasi';
+        $app = ListApp::where('slug', $slug)->firstOrFail();
+        return view('pages.listapp.view', compact('app', 'apps_menu', 'apps_submenu'));
     }
 
-    public function update(Request $request, $id)
+    public function edit($slug)
+    {
+        $apps_menu = 'Kelola Aplikasi';
+        $apps_submenu = 'Edit Aplikasi';
+        $app = ListApp::where('slug', $slug)->firstOrFail();
+        return view('pages.listapp.edit', compact('app', 'apps_menu', 'apps_submenu'));
+    }
+
+    public function update(Request $request, $slug)
     {
         $request->validate([
             'nama' => 'required|string|max:255',
@@ -77,7 +96,7 @@ class ListAppController extends Controller
             'logo' => 'mime:jpeg,png,jpg,gif,svg|max:2048',
         ]);
 
-        $app = ListApp::findOrFail($id);
+        $app = ListApp::where('slug', $slug)->firstOrFail();
         $app->nama = $request->nama;
         $app->deskripsi = $request->deskripsi;
         $app->akses = $request->akses;
@@ -90,18 +109,34 @@ class ListAppController extends Controller
             $image = time() . '_' . $file->getClientOriginalName();
             $file->move(public_path('/img'), $image);
             $app->logo = $image;
+        } else {
+            $app->logo = 'logo_bps.png';
         }
 
         $app->save();
-        return redirect()->route('dashboard.index')->with('success', 'Aplikasi berhasil diperbarui.');
+        Swal::fire([
+            'title' => 'Informasi',
+            'text' => 'Aplikasi ' . $app->nama . ' berhasil diperbaharui!',
+            'icon' => 'success',
+            'confirmButtonText' => 'Tutup'
+        ]);
+        return redirect()->route('listapp.index');
     }
 
-    public function delete($id)
+    public function destroy($slug)
     {
-        $app = ListApp::findOrFail($id);
-        File::delete(public_path('/img/' . $app->logo));
+        $app = ListApp::where('slug', $slug)->firstOrFail();
+        if ($app->logo != 'logo_bps.png') {
+            File::delete(public_path('/img/' . $app->logo));
+        }
 
         $app->delete();
-        return redirect()->route('dashboard.index')->with('success', 'Aplikasi berhasil dihapus.');
+        Swal::fire([
+            'title' => 'Informasi',
+            'text' => 'Aplikasi ' . $app->nama . ' berhasil dihapus!',
+            'icon' => 'success',
+            'confirmButtonText' => 'Tutup'
+        ]);
+        return redirect()->route('listapp.index');
     }
 }
